+++ attributes/cert.rb
@@ -2,6 +2,8 @@
 
 tcb = 'http_platform'
 
+default[tcb]['cert']['owner_group'] = 'root'
+
 default[tcb]['cert']['expiration_days'] = 365
 default[tcb]['cert']['rsa_bits'] = 2048
 default[tcb]['cert']['dh_param']['bits'] = 2048