+++ recipes/vault_cert.rb
@@ -9,6 +9,9 @@ key = node[tcb]['cert']['vault_item_key']
 file path_to_vault_cert do
   # must be lazy because hostname may change!
   content lazy { vault_secret(bag, item, key) } # rubocop:disable Lint/AmbiguousBlockAssociation
+  owner 'root'
+  group node[tcb]['cert']['owner_group']
+  mode '0644'
 end
 
 key_key = node[TCB]['key']['vault_item_key']
@@ -18,7 +21,7 @@ file path_to_vault_key do
   content lazy { vault_secret(bag, item, key_key) } # rubocop:disable Lint/AmbiguousBlockAssociation
   sensitive true
   owner 'root'
-  group 'root'
-  mode '0600'
+  group node[tcb]['cert']['owner_group']
+  mode '0640'
   only_if { key_key }
 end