+++ test/integration/default/cert_test.rb
@@ -40,18 +40,18 @@ end
 describe file(path_to_self_signed_cert(node)) do
   it { should exist }
   it { should be_file }
-  it { should be_mode 0o600 }
+  it { should be_mode 0o640 }
   it { should be_owned_by 'root' }
-  it { should be_grouped_into 'root' }
+  it { should be_grouped_into 'adm' }
   its(:content) { should match 'BEGIN CERTIFICATE' }
 end
 
 describe file(path_to_self_signed_key(node)) do
   it { should exist }
   it { should be_file }
-  it { should be_mode 0o600 }
+  it { should be_mode 0o640 }
   it { should be_owned_by 'root' }
-  it { should be_grouped_into 'root' }
+  it { should be_grouped_into 'adm' }
   its(:content) { should match 'BEGIN RSA PRIVATE KEY' }
 end
 
@@ -108,8 +108,8 @@ end
 describe file(File.join(cert_public_dir(node), 'http_platform_csr.pem')) do
   it { should exist }
   it { should be_file }
-  it { should be_mode 0o600 }
+  it { should be_mode 0o644 }
   it { should be_owned_by 'root' }
-  it { should be_grouped_into 'root' }
+  it { should be_grouped_into 'adm' }
   its(:content) { should match 'BEGIN CERTIFICATE REQUEST' }
 end