+++ test/integration/lets_encrypt_cert/lets_encrypt_test.rb
@@ -22,6 +22,14 @@ describe file(path_to_lets_encrypt_key) do
   its(:content) { should match 'BEGIN RSA PRIVATE KEY' }
 end
 
+describe file('/etc/letsencrypt/live') do
+  it { should exist }
+  it { should be_directory }
+  it { should be_mode 0o750 }
+  it { should be_owned_by 'root' }
+  it { should be_grouped_into 'adm' }
+end
+
 # Special fields for this cert
 describe file(path_to_ssl_host_conf(node)) do
   its(:content) { should match "SSLCertificateFile #{path_to_lets_encrypt_cert}" }